+++ database/migrations/2017_02_16_205125_create_users_table.php
@@ -39,6 +39,7 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('users');
+		if(Schema::hasTable('users'))
+        	Schema::drop('users');
     }
 }
