+++ database/migrations/2017_02_16_205126_create_password_resets_table.php
@@ -27,6 +27,7 @@ class CreatePasswordResetsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('password_resets');
+		if(Schema::hasTable('password_resets'))
+        	Schema::drop('password_resets');
     }
 }
