+++ database/migrations/2017_02_16_205313_task_logs.php
@@ -31,6 +31,7 @@ class TaskLogs extends Migration
      */
     public function down()
     {
-        Schema::drop('task_logs');
+		if(Schema::hasTable('task_logs'))
+        	Schema::drop('task_logs');
     }
 }
